+++ src/Filament/GadyaCmsPlugin.php
@@ -13,6 +13,7 @@ use Gadya\Cms\Filament\Pages\ArticleGenerator;
 use Gadya\Cms\Filament\Pages\Blocks;
 use Gadya\Cms\Filament\Pages\Dashboard;
 use Gadya\Cms\Filament\Pages\Emails;
+use Gadya\Cms\Filament\Pages\GetFound;
 use Gadya\Cms\Filament\Pages\Globals;
 use Gadya\Cms\Filament\Pages\Navigation;
 use Gadya\Cms\Filament\Pages\SearchSettings;
@@ -338,6 +339,7 @@ class GadyaCmsPlugin implements Plugin
                 Navigation::class,
                 $this->hasAi() ? AiSettings::class : null,
                 $this->hasSearch() ? SearchSettings::class : null,
+                GetFound::class,
                 SiteStatus::class,
                 $this->hasForms() ? Emails::class : null,
                 $this->hasAi() && $this->hasBlog() ? ArticleGenerator::class : null,